requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sentence-transformers>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0
